use module augmentation as part of the plugin
