+++ package.json
@@ -49,15 +49,18 @@
 		"ts-loader": "^9.4.1",
 		"ts-node": "^10.9.1",
 		"typescript": "^4.7.4",
-		"web3-core": "5.0.0",
-		"web3-eth": "5.0.0",
-		"web3-eth-abi": "5.0.0",
-		"web3-eth-contract": "5.0.0",
-		"web3-types": "5.0.0",
-		"web3-validator": "5.0.0"
+		"web3": "4.0.1-alpha.1",
+		"web3-core": "4.0.1-alpha.1",
+		"web3-eth": "4.0.1-alpha.1",
+		"web3-eth-abi": "4.0.1-alpha.1",
+		"web3-eth-contract": "4.0.1-alpha.1",
+		"web3-types": "0.1.1-alpha.1",
+		"web3-validator": "0.1.1-alpha.1"
 	},
 	"peerDependencies": {
+		"web3": ">= 4.0.1-alpha.0 < 5",
 		"web3-core": ">= 4.0.1-alpha.0 < 5",
+		"web3-eth": ">= 4.0.1-alpha.0 < 5",
 		"web3-eth-abi": ">= 4.0.1-alpha.0 < 5",
 		"web3-eth-contract": ">= 4.0.1-alpha.0 < 5",
 		"web3-types": ">= 0.1.1-alpha.0 < 5",
