+++ src/reexported_web3_objects.ts
@@ -1,4 +1,5 @@
+import Web3 from 'web3';
 import { Web3Context } from 'web3-core';
 import Web3Eth from 'web3-eth';
 
-export { Web3Context, Web3Eth };
+export { Web3, Web3Context, Web3Eth };
